+++ mycodo/inputs/ds1825.py
@@ -59,9 +59,10 @@ class InputModule(AbstractInput):
 
     def initialize_input(self):
         from w1thermsensor import W1ThermSensor
+        from w1thermsensor import Sensor
 
         self.sensor = W1ThermSensor(
-            W1ThermSensor.THERM_SENSOR_DS1825, self.input_dev.location)
+            Sensor.DS1825, self.input_dev.location)
         if self.input_dev.resolution:
             self.sensor.set_resolution(self.input_dev.resolution)